pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.8.0,<3.9"
numpy = "^1.22.2"
python-dotenv = "^0.19.2"
discord = "^1.7.3"
wit = "^6.0.0"
requests = "^2.27.1"
pandas = "^1.4.1"
recommendation = "^0.3.1"
scipy = "^1.8.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"